It now resembles a library a little more
1. Separated logic off graphics in Sketch
2. Class Ship moved over to the logic part
3. Added parametrized sketch and just logic instantiation
4. Renamed some functions and events
5. Removed that shenanigan with getting ejs data into javascript
variable via creating a p tag, dumping the ejs data into there and
then reading it out. I rejected to work with ejs, because data can be
soaked dynamically, via socket.io, which I'm in love with.
6. Most of the code refactored and a little optimized.
7. I use Lodash just a bit, but that is not really necessary. Now that
is used just to create 10 by 10 arrays

In plans:

1. Random ship arrangement
2. Fair different board dimensions
3. Different schema objects
4. AI with little code that would aim to finish off injured ships but
shoot randomly if when are none.
5. Make a more advanced AI using tf.js and neural nets.
6. Ways to integrate those in the game.