Bot
Added Random Bot. It finishes off ships if there are injured ones and
otherwise shoots randomly.

Added way to arrange ships automatically (randomly).

Coming up: Evolutionary NN that learns to play this game. Probably will
be done in a separate project. Here would be only the values of the
weights and the structure of the NN, so that any user could play against
that bot without the server.

Also some minor and debug changes.